requirements.txt
Django
Flake8
Pytest-django
python-decouple
dj-database-url
psycopg2
djangorestframework